--- src/main/kotlin/lottery/domain/LotteryGenerator.kt
@@ -1,5 +1,0 @@
-package lottery.domain
-
-interface LotteryGenerator {
-    fun generate(): Lottery
-}
