+++ src/main/kotlin/lottery/domain/lottery/Lotteries.kt
@@ -1,5 +1,6 @@
-package lottery.domain
+package lottery.domain.lottery
 
+import lottery.domain.Rank
 import lottery.domain.Rank.Companion.fillMissRankWithDefault
 
 class Lotteries(
