+++ src/main/kotlin/lottery/domain/lottery/LotteryNumber.kt
@@ -1,4 +1,4 @@
-package lottery.domain
+package lottery.domain.lottery
 
 @JvmInline
 value class LotteryNumber(
