+++ src/main/kotlin/lottery/domain/lottery/generator/LotteryGenerator.kt
@@ -1,0 +1,7 @@
+package lottery.domain.lottery.generator
+
+import lottery.domain.lottery.Lottery
+
+interface LotteryGenerator {
+    fun generate(): Lottery
+}
