+++ src/test/kotlin/lottery/domain/MoneyTest.kt
@@ -4,6 +4,7 @@ import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
+import lottery.domain.lottery.generator.RandomLotteryGenerator
 
 class MoneyTest : FunSpec({
 
@@ -16,7 +17,9 @@ class MoneyTest : FunSpec({
 
     context("purchaseLotteries") {
         test("로또를 사기에 부족한 돈이 입력될 경우 예외가 발생한다") {
-            val exception = shouldThrowExactly<IllegalStateException> { Money(value = 999).purchaseLotteries(RandomLotteryGenerator) }
+            val exception = shouldThrowExactly<IllegalStateException> { Money(value = 999).purchaseLotteries(
+                RandomLotteryGenerator
+            ) }
             exception.message shouldBe "로또를 사기엔 부족한 금액이다"
         }
