+++ src/test/kotlin/lottery/domain/lottery/LotteriesTest.kt
@@ -1,13 +1,14 @@
-package lottery.domain
+package lottery.domain.lottery
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldContainAll
 import io.kotest.matchers.shouldBe
-import lottery.domain.Lotteries.Companion.toLotteries
-import lottery.domain.LotteryTest.Companion.LOTTERY_1_6
-import lottery.domain.LotteryTest.Companion.LOTTERY_2_7
-import lottery.domain.LotteryTest.Companion.LOTTERY_3_8
-import lottery.domain.LotteryTest.Companion.LOTTERY_4_9
+import lottery.domain.Rank
+import lottery.domain.lottery.Lotteries.Companion.toLotteries
+import lottery.domain.lottery.LotteryTest.Companion.LOTTERY_1_6
+import lottery.domain.lottery.LotteryTest.Companion.LOTTERY_2_7
+import lottery.domain.lottery.LotteryTest.Companion.LOTTERY_3_8
+import lottery.domain.lottery.LotteryTest.Companion.LOTTERY_4_9
 
 class LotteriesTest : FunSpec({
