+++ src/test/kotlin/lottery/domain/lottery/LotteryTest.kt
@@ -1,4 +1,4 @@
-package lottery.domain
+package lottery.domain.lottery
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
@@ -6,15 +6,16 @@ import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_1
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_2
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_3
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_4
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_5
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_6
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_7
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_8
-import lottery.domain.LotteryNumberTest.Companion.LOTTERY_NUMBER_9
+import lottery.domain.Rank
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_1
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_2
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_3
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_4
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_5
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_6
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_7
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_8
+import lottery.domain.lottery.LotteryNumberTest.Companion.LOTTERY_NUMBER_9
 
 class LotteryTest : FunSpec({
